move the edit homepage related files to settings folder
