+++ datahub-web-react/src/app/homeV3/HomePageContent.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 import { useGlobalSettings } from '@app/context/GlobalSettingsContext';
 import { useUserContext } from '@app/context/useUserContext';
-import EditDefaultTemplateBar from '@app/homeV3/EditDefaultTemplateBar';
-import EditDefaultTemplateButton from '@app/homeV3/EditDefaultTemplateButton';
 import { Announcements } from '@app/homeV3/announcements/Announcements';
 import { PageTemplateProvider } from '@app/homeV3/context/PageTemplateContext';
+import EditDefaultTemplateBar from '@app/homeV3/settings/EditDefaultTemplateBar';
+import EditDefaultTemplateButton from '@app/homeV3/settings/EditDefaultTemplateButton';
 import { CenteredContainer, ContentContainer, ContentDiv } from '@app/homeV3/styledComponents';
 import Template from '@app/homeV3/template/Template';
